requirements.txt
requests
bs4
pandas
demjson
progressbar2
